requirements.txt
langchain
faiss-cpu
sentence-transformers
torch
transformers
pypdf
python-docx
pandas
openpyxl
beautifulsoup4
markdown
unstructured
pydantic
pydantic-settings
python-dotenv
plotly
pandas
fpdf2
tabula-py
openpyxl
python-docx
pytest
pytest-cov
plotly
fpdf2
pandas
tabula-py